+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.26",
-  "contentSha256_12": "7ee2ea326278",
+  "pkgVersion": "1.1.0-alpha.27",
+  "contentSha256_12": "e9f09aff0ce9",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.0-alpha.26.js"
+  "hacsFile": "nodalia-cards-1.1.0-alpha.27.js"
 };
-export const pkgVersion = "1.1.0-alpha.26";
-export const contentSha256_12 = "7ee2ea326278";
+export const pkgVersion = "1.1.0-alpha.27";
+export const contentSha256_12 = "e9f09aff0ce9";
 export const file = "nodalia-cards.bundle.js";